Added dynamic loading of content
Using v-for loops to generate <li> items to generate all the nav <li> items upon loading.

create another vue app in the bottom script tag.

Also used string intepolation to dynamically load title of the <a> tag of the <li> item.

index.js is just a test I did to add modularity but it didnt as planned.